compilation error
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@2703 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/project/Project.java
@@ -1723,8 +1723,8 @@ public class Project extends Listener
 			if (retrieve)
 			{
 				int index = table.getSelectedRow();
-				Integer version = TextUtils.atoi((String)dataModel.getValueAt(index, 0));
-				new GetOldVersionJob(cell, version.intValue());
+				int version = TextUtils.atoi((String)dataModel.getValueAt(index, 0));
+				new GetOldVersionJob(cell, version);
 			} else
 			{
 				dispose();
